Squashed 'cfsat/' changes from ddeb335..c8e773a
c8e773a Merge commit '4a8c2b4ab2ef1b2514e99316a06ba9476ad06cfe' as 'osk-apps/file_xfer'
4a8c2b4 Squashed 'osk-apps/file_xfer/' content from commit e003cb0
ae49b62 Replacing subtree after cfs-apps renamed to osk-apps
4e6ea0f Merge commit '62fa0dd304c2ddae9504104e899c440ac2075760' as 'osk-apps/file_mgr'
62fa0dd Squashed 'osk-apps/file_mgr/' content from commit b187128
10c9e8f Replacing subtree after cfs-apps renamed to osk-apps
07a2933 Merge commit '460bb3eb1535ee18de37da392d9a875839c96bc5' as 'osk-apps/osk_c_demo'
460bb3e Squashed 'osk-apps/osk_c_demo/' content from commit 8ef3c1f
68e6d06 Replacing subtree after cfs-apps renamed to osk-apps
d575725 Merge commit '1b9010948ca32cf6302fe9265e6aada534641364' as 'osk-apps/osk_c_fw'
1b90109 Squashed 'osk-apps/osk_c_fw/' content from commit c16730c
299da96 Replacing subtree after cfs-apps renamed to osk-apps
ac715d5 Fixed detector image alignment

git-subtree-dir: cfsat
git-subtree-split: c8e773a9833ab3f8c6c67befb1e95d01cb38b4f1

+++ osk-apps/file_mgr/fsw/src/dir.c
@@ -645,7 +645,7 @@ bool SendDirListTlm(const char *DirName, uint16 DirListOffset, bool IncludeSizeT
    char   PathFilename[OS_MAX_PATH_LEN] = "\0";
 
 
-   OS_printf("sizeof(Dir->ListTlm.Payload.FileList) = %ld\n", sizeof(Dir->ListTlm.Payload.FileList));
+   OS_printf("sizeof(Dir->ListTlm.Payload.FileList) = %d\n", (unsigned int)sizeof(Dir->ListTlm.Payload.FileList));
 
    FileInfo = FileUtil_GetFileInfo(DirName, OS_MAX_PATH_LEN, false);
